Add benchmark to Gemfile. Use benchmark to measure the time Page.search queries take. Set the benchmark time to an erb attribute so it can be displayed in the view.
Refactor Page.search to include a skip parameter. The skip parameter is used within the mongoid find call.
Refactor the pages/search controller url to include a page number. Use the page number to determine how many records to skip when the Page.search method is called. Add forms to the page/search view for pagination.